fix: move feed file to feature folder
