+++ src/features/feed/ChallengeRoomFeedList.tsx
@@ -5,9 +5,10 @@ import { useScrollToBottom } from '@/shared/hooks';
 import { isFeedDateDifferent } from '@/shared/utils/date';
 import { TChallengeFeed } from '@/types/feed';
 
-import { DateChip } from '../chip';
-import { MyFeed } from '../feed/MyFeed';
-import { OthersFeed } from '../feed/OthersFeed';
+import { DateChip } from '../../my-poor-room/components/chip';
+
+import { MyFeed } from './MyFeed';
+import { OthersFeed } from './OthersFeed';
 
 // MOCK
 const challengeFeedList: TChallengeFeed[] = [
