+++ src/features/feed/index.ts
@@ -1,0 +1,6 @@
+import { ChallengeRoomFeedList } from './ChallengeRoomFeedList';
+import { MyFeed } from './MyFeed';
+import { MyRoomFeedList } from './MyRoomFeedList';
+import { OthersFeed } from './OthersFeed';
+
+export { MyFeed, OthersFeed, MyRoomFeedList, ChallengeRoomFeedList };
