--- src/my-poor-room/components/feed-list/index.ts
@@ -1,4 +1,0 @@
-import { ChallengeRoomFeedList } from './ChallengeRoomFeedList';
-import { MyRoomFeedList } from './MyRoomFeedList';
-
-export { MyRoomFeedList, ChallengeRoomFeedList };
