--- src/my-poor-room/components/feed/index.ts
@@ -1,4 +1,0 @@
-import { MyFeed } from './MyFeed';
-import { OthersFeed } from './OthersFeed';
-
-export { MyFeed, OthersFeed };
