+++ src/pages/my-poor-room/index.tsx
@@ -1,12 +1,9 @@
 import { useRouter } from 'next/router';
 import { ReactElement } from 'react';
 
+import { ChallengeRoomFeedList, MyRoomFeedList } from '@/features/feed';
 import { ChallengeAchievement } from '@/my-poor-room/components/challenge-achievement';
 import { ChallengeCategories } from '@/my-poor-room/components/challenge-categories';
-import {
-  ChallengeRoomFeedList,
-  MyRoomFeedList,
-} from '@/my-poor-room/components/feed-list';
 import { Spacing } from '@/shared/components';
 import BottomNavLayout from '@/shared/components/layout/BottomNavLayout';
